requirements2 2.txt
fastapi
uvicorn
scikit-learn
joblib
python-dotenv
python-multipart
python-magic
requests
gunicorn